Fix CmdHaddock to follow CmdBuild pattern rather than CmdConfigure
It looks like CmdHaddock was based on the CmdConfigure template and so
copied some bits that don't make sense for it. Change those bits to be
the same as for build.

In particular for printing the plan, the configure command always
behaves as if it is in --dry-run mode, but that's not appropriate for
the haddock command. The other change is to follow the standard pattern
for the runProjectBuildPhase / runProjectPostBuildPhase and the handling
of the --dry-run option, which is handled internally by those and does
not need to be handled explicitly in CmdHaddock.