Rename BuildTarget to TargetSelector
We're going to generalise it and the name will no longer be appropriate.
We'll add the ability to select all packages, or to select only certain
kinds of components in a package. So this makes it less like a resolved
target and more or an expression that needs further work to resolve to
a set of targets.