Split getting project config from making the install plan
Previously we had a rebuildInstallPlan that does everything up to
producing the elaborated install plan. This included reading the project
config and all the local package description files etc.

It is useful to split this latter aspect out separately. The reason for
this is that we would really prefer to resolve command line build
targets without first having to run the solver and produce the
elaborated install plan. We would prefer to be able to report mistakes
in the build targets more or less instantly rather than having to wait
many seconds (or sometime more) before being able to report problems.
Worse, with the current scheme the solver may fail to find a solution
and so we'll end up reporting that rather than the problem with the
user's build target. Indeed problems with a build target are most likely
to happen when the project config is messed up (or not what users
expect, or before they understand the model) and those are the cases
where it's most likely that there is no project solution.

So this is a step towards being able to do all target resolution prior
to solving.