Split early steps out of runProjectPreBuildPhase
This is part of the refactoring to resolve command line build targets
without first having to run the solver and produce the elaborated
install plan.

Previously runProjectPreBuildPhase would do all the early steps
including finding the project root and establishing the project
config and the local packages. This patch splits those early steps
out into a separate pass in the project orchestration. It also
eliminates the need for the hookPrePlanning and since there's then
only one hook left we don't bother to name the PreBuildHooks.