Consolodate error reporting in readTargetSelectors
So there's now just one error type rather than two, and errors are
returned rather than reported directly.

Now that we have squashed together the two phases of the target
selector parsing and resolving, we also want to do the same with the
error reporting for the two phases. And to make it possible to customise
the error messages for different commands we have the reporting being
done explicitly and separately.