More regular fully qualified syntax for target selectors
The command line target syntax has two main use cases: humans who prefer
short forms and can deal with occasional ambiguity, and programs/scripts
which need unambiguous syntax but it's ok if it's a big long.

This patch changes the syntax for the long unambiguous form. It makes
sure we have a fully qualified form for each target kind, and makes them
consistent with each other. This is the syntax that will need to be
documented for script authors (possibly in the --help in addition to the
user guide).