Make module name targets shadow file name targets
Generally these two are not ambigious with each other, e.g. "P" vs
"P.hs" or "Data.P" vs "Data/P", but since we allow matching file names
independent of file extension then for single-component module names
then we have an ambiguity between module "P" and file name root "P". But
these of course refer to the same thing, so the ambiguity is not useful.
So we shadow these cases so it refers to the module name.