use the install plan to modify PATH in new-exec
Previously, cabal new-exec would modify the PATH by including each
package's bin directory if it contained executables. This has some
drawbacks:

* Lots of filesystem probing
* Can have false positives due to dynamic libraries
* Mildly unpredictable

The new plan is to add each package's bin directory if the package
includes an executable stanza. This may include some directories that
have no executables in them (e.g. if the executable is not yet built or
is not in the install plan), but is more reproducible and requires less
filesystem access.