allow new-exec'ing of executables built inplace
Previously, new-exec would use the `build` directory when looking for
executables built inplace. But these executables are actually in
subdirectories of `build`, so the search would fail (potentially falling
back to an executable in the user's PATH instead). Now we place one
directory in the PATH for each executable in each inplace-build package
in the install plan.