Adjust args to selectComponentTarget functions
Previously they took a TargetSelector that would always be a
TargetComponent constructor, which meant that if we were interested in
the content then we'd have to do partial pattern matching. Instead they
now take all the contents of the TargetComponent as separate args which
means no partial patterns and easy access to the parts.

Also where relevant, put the TargetComponent contents into the
TargetProblemCommon and TargetProblem constructors.