Rewrite and document the selectPackageTargets for each command
This is the function that implements the semantics of the target
selector matching for each command.

Introduce and use utils for target filtering and selection.

Rename several of the error constructors to be consistent. For the test,
bench and run commands, add a distinction between there being no targets
at all and none of the right kind.

Document the selectPackageTargets top level and cases, and
selectComponentTarget. Also document the TargetProblem and constructors.