Adjust multiple target error case for run and repl commands
The run and repl commands only work with a single component as targets.
When the user gives multiple target selectors that refer to distinct
targets then we must report an error. It turns out that the information
we already had was not really sufficient to construct a helpful error
message. The information we had was the component targets that were
actually selected, but we actually want to know the target selectors
that gave rise to the selection of multiple targets.

So this patch extends the information we collect in the resolveTargets
function and stashes this info in the TargetProblemMultipleTargets
constructor, which will enable construction of useful error messages.