Pass other-modules to `ghc --make` for exe components
`gbuildSources` was introduced by a refactoring in
b2a2ae9ed5f035d8dcaf24daf9886645b653fcfc but there
doesn't appear to be a good reason why the modules are
only passed for a non-haskell main module.

So this patch consistently passes `other-modules` to
`ghc --make` for both variants of executables.

Addresses part of #4528