Refactor handling of conflict set representing option to not choose a goal.
The conflict set that represents the option to not choose a value for a goal,
'initial', has two purposes:

1. 'initial' contributes to a node's conflict set, because it is unioned with
   the other conflict sets under the same choice node, if the current variable
   is contained in all of them.
2. 'initial' contributes to the conflict count.  When it is unioned with the
   other conflict sets, it is also added to the ConflictMap.

Previously, 'initial' was treated as the first of a node's children for (1) and
the last of the children for (2).  This refactoring treats 'initial' as the last
child in both cases, to make the code clearer, and renames it to 'lastCS'.
There should be no change in behavior.