Remove polymorphism from runProjectPreBuildPhase
Return a monomorphic TargetsMap from selectPlanSubset and put it in the
final ProjectBuildContext.

This causes slight duplication in CmdRun.hs, but not as bas as when we
didn't return anything. In exchange, the final return type is much
simpler.

Note that CmdConfigure returns an empty TargetsMap, as configure doesn't
accept targets.