Don't paramaterise the TargetSelector type
Previously the TargetSelector type had a type param for the type of the
package that it referred to. In particular we used it with types like:

type Matcher  = ... -> Match (TargetSelector KnownPackage)
type Renderer = TargetSelector PackageId -> ...

However we are about to extend the TargetSelector so that it does not
just refer to one form of package (e.g. KnownPackage) but can refer to
packages via various different forms and partial information. So it no
longer makes sense to have TargetSelector be paramaterised by the
different states of the one kind of package it refers to, as there are
now many kinds. So in preparation for that we simplify it so that it is
equivalent to always using TargetSelector PackageId, and we remove the
type paramater.