Solver: Use more specific failure reasons for dependency conflicts.
This commit splits the 'Conflicting' constructor of 'FailReason' into five
different constructors, for different types of dependency conflicts.  This
change makes it easier to print different error messages for different
conflicts, though it doesn't change any of the error messages yet.

The refactoring caused a small change to the error message for conflicts caused
by self-dependencies.