Change order of hsLibraries
After using this for a while. It occured to me that due to linker ordering, and GHC just forwarding the libraries as specified in the `hsLibraries` field, we should put the `extraBundledLibs` at the end. As such the haskell library can reference the `extraBundledLibs`.

This still leave the question of recursively linked libs. But getting that linked right should rather be deal with in GHC, which optimally would use being and end groups for the libraries to link.